requirements.txt
python-dotenv==1.0.1
streamlit==1.35.0
faiss-cpu==1.7.4
sentence-transformers==2.6.1
langchain==0.2.1
langchain-community==0.2.1
google-generativeai==0.5.4
tqdm==4.66.4
numpy==1.26.4
nltk==3.8.1
pypdf==3.17.1